backend/nash/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3

channels>=4.0
channels-redis>=4.1

daphne>=4.0
gunicorn>=21.2

psycopg2-binary>=2.9

redis>=5.0

whitenoise>=6.6

python-dotenv>=1.0

Pillow>=10.0
